break out API methods into separate classes
