+++ webapp/src/api/LoginApi.ts
@@ -1,5 +1,7 @@
+import axios from "axios";
 import { Buffer } from "buffer";
-import { attemptLogin } from "./ExchangeApi";
+import { APPLICATION_JSON } from "../utils/StringConstants";
+import { loginUrl } from "../utils/Urls";
 
 export const login = async (username: string, password: string) => {
   const jsonPayload = {
@@ -14,4 +16,21 @@ export const login = async (username: string, password: string) => {
     console.error("Error on login.  Message: " + e.message);
     return null;
   }
+};
+
+export const attemptLogin = async (jsonData: {
+  username: string;
+  password: string;
+}) => {
+  const res = await axios({
+    timeout: 300000,
+    url: loginUrl(),
+    method: "POST",
+    data: jsonData,
+    headers: {
+      "Content-Type": APPLICATION_JSON,
+    },
+  });
+
+  return res;
 };
