+++ webapp/src/screens/tabs/UsersTable.tsx
@@ -10,11 +10,6 @@ import Paper from '@mui/material/Paper';
 import Toolbar from "@mui/material/Toolbar";
 import {ChangeEvent, useCallback, useEffect, useMemo, useState} from "react";
 import {
-  bulkUploadUsers,
-  createNewUser,
-  deleteUser,
-  getAllUsers} from "../../api/ExchangeApi";
-import {
   Alert,
   alpha,
   Checkbox,
@@ -44,6 +39,7 @@ import {
 } from "../../utils/ExcelUtils";
 import xlsx from "json-as-xlsx";
 import DownloadIcon from '@mui/icons-material/Download';
+import { bulkUploadUsers, createNewUser, deleteUser, getAllUsers } from '../../api/UserApi';
 
 export interface UserDoc {
   id: string,
